requirements/dev.txt
-r common.txt
django-debug-toolbar==1.8
django-decouple==2.1
sqlparse==0.2.4

